+++ railties/test/generators/generators_test_helper.rb
@@ -1,5 +1,6 @@
 require 'abstract_unit'
 require 'active_support/core_ext/module/remove_method'
+require 'active_support/testing/stream'
 require 'rails/generators'
 require 'rails/generators/test_case'
 
@@ -23,6 +24,8 @@ require 'action_dispatch'
 require 'action_view'
 
 module GeneratorsTestHelper
+  include ActiveSupport::Testing::Stream
+
   def self.included(base)
     base.class_eval do
       destination File.join(Rails.root, "tmp")
@@ -42,21 +45,4 @@ module GeneratorsTestHelper
     FileUtils.cp routes, destination
   end
 
-  def quietly
-    silence_stream(STDOUT) do
-      silence_stream(STDERR) do
-        yield
-      end
-    end
-  end
-
-  def silence_stream(stream)
-    old_stream = stream.dup
-    stream.reopen(IO::NULL)
-    stream.sync = true
-    yield
-  ensure
-    stream.reopen(old_stream)
-    old_stream.close
-  end
 end
